pyproject.toml
[project]
name = "opsenv"
version = "1.0.0"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "requests>=2.31.0",
  "pydantic>=2.0.0",
  "openai>=1.0.0",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
  "pyyaml>=6.0",
]